+++ libs/langgraph/.gitignore
@@ -1,7 +1,3 @@
-_types.cjs
-_types.js
-_types.d.ts
-_types.d.cts
 index.cjs
 index.js
 index.d.ts
