+++ libs/langgraph/langchain.config.js
@@ -12,7 +12,6 @@ function abs(relativePath) {
 export const config = {
   internals: [/node\:/, /@langchain\/core\//, /async_hooks/],
   entrypoints: {
-    _types: "_types",
     index: "index",
     web: "web",
     pregel: "pregel/index",
