+++ libs/langgraph/package.json
@@ -81,15 +81,6 @@
     "registry": "https://registry.npmjs.org/"
   },
   "exports": {
-    "./_types": {
-      "types": {
-        "import": "./_types.d.ts",
-        "require": "./_types.d.cts",
-        "default": "./_types.d.ts"
-      },
-      "import": "./_types.js",
-      "require": "./_types.cjs"
-    },
     ".": {
       "types": {
         "import": "./index.d.ts",
@@ -130,10 +121,6 @@
   },
   "files": [
     "dist/",
-    "_types.cjs",
-    "_types.js",
-    "_types.d.ts",
-    "_types.d.cts",
     "index.cjs",
     "index.js",
     "index.d.ts",
