--- libs/langgraph/src/_types.ts
@@ -1,11 +1,0 @@
-// Export types for portability, these shouldn't be needed/used directly
-export {
-  type AnyValue,
-  type WaitForNames,
-  type DynamicBarrierValue,
-  type LastValue,
-  type NamedBarrierValue,
-  type Topic,
-} from "./channels/index.js";
-
-export { type AnnotationRoot } from "./graph/index.js";
