+++ libs/langgraph/src/web.ts
@@ -21,7 +21,14 @@ export {
   BaseChannel,
   type BinaryOperator,
   BinaryOperatorAggregate,
+  type AnyValue,
+  type WaitForNames,
+  type DynamicBarrierValue,
+  type LastValue,
+  type NamedBarrierValue,
+  type Topic,
 } from "./channels/index.js";
+export { type AnnotationRoot as _INTERNAL_ANNOTATION_ROOT } from "./graph/index.js";
 export { type RetryPolicy } from "./pregel/utils.js";
 export { Send } from "./constants.js";
